% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytobandToArm.R
\name{cytobandToArm}
\alias{cytobandToArm}
\title{cytobandToArm: Reduces cytoband data to chromosome arms}
\usage{
cytobandToArm(cytoband)
}
\arguments{
\item{cytoband}{cytoband data from data(ucsc.[GENOMEBUILD].cytoband)}
}
\value{
List of chromosome arms and their arm-level coordinates in dataframes
}
\description{
Takes the data([CYTOBAND]) information and splits the 
cytobands into their individual p-arm, q-arm or centromere positions.
Then it will reduce and clean up the data before returning it as a list
of chromosomal positions with 3 rows: 'p', 'q', and 'cen'.
}
\examples{
data(ucsc.hg19.cytoband)
cytobandToArm(ucsc.hg19.cytoband)
}
