% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.classifyCN}
\alias{.classifyCN}
\title{Classify CN into Loss/Neut/Gain}
\usage{
.classifyCN(
  cn,
  ploidy = 2,
  threshold = 0.5,
  neg_max = -100,
  pos_max = 100,
  verbose = FALSE
)
}
\arguments{
\item{cn}{A vector of CN values [Numeric]}

\item{ploidy}{A numeric value for base ploidy (Default=2) [Numeric]}

\item{threshold}{A numeric value for the threshold around the 
ploidy to be considered (e.g. ploidy +/- threshold) (Default=0.5) [Numeric]}

\item{neg_max}{The maximum negative value (Default=-100)}

\item{pos_max}{The maximum positive value (Default=100)}

\item{verbose}{Print warning statements or not}
}
\description{
Classifies CN aberrations as either a Loss (-1), 
neutral (0), or gain (1) relative to some ploidy number +/-
a threshold
}
