% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.getChrarmFractions}
\alias{.getChrarmFractions}
\title{Gets fraction of intervals that make up each GR object}
\usage{
.getChrarmFractions(gr, cn_col = "CN")
}
\arguments{
\item{gr}{GRanges object of a chromosome or chromosome arm}

\item{cn_col}{Column in GRanges object containing CN values [Default='CN']}
}
\value{
2-column data frame of chr_fractions with NA's included 
and NA's excluded
}
\description{
Takes a GRanges object (ideally and designed for
a single chromosome or chromosome arm) and founds out what fraction
of the total chr/chr_arm each interval makes up of it. It does this
for all intervals with a copy number value (non-NA in cn_col), as 
well as with NA's removed (no CN value (NA) in cn_col).
}
