% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.validateSeg}
\alias{.validateSeg}
\title{Validates if the seg is compatible with this R package}
\usage{
.validateSeg(seg)
}
\arguments{
\item{seg}{a Seg file dataframe}
}
\value{
LOGICAL
}
\description{
Validates if the seg is compatible with this R package
}
\examples{
#.validateSeg(data.frame("Chrom"=1, "start"=1, "end"=2))
#.validateSeg(data.frame("chr"=1, "X"=1, "loc.end"=2))
}
