% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceArms.R
\name{reduceArms}
\alias{reduceArms}
\title{Reduce Arm/Seg level CN to an aneuploidy score}
\usage{
reduceArms(
  segf_caa,
  caa_method = c("arm", "seg"),
  threshold = NA,
  arm_ids = c("p", "q"),
  verbose = FALSE
)
}
\arguments{
\item{segf_caa}{Output from getCAA(), GRanges of individual 
chromosomes [GRangesList]}

\item{caa_method}{The chromosomal-arm aneuploidy scoring metric to use: [Vector]
- 'arm': Reduces the arm-level CN classifications (classified from the Weighted Median of the chr-arm CN) to a single value 
- 'seg': Uses a threshold (or weighted median if no threshold given) on the chr-arm CN classifications across all arm-level segs}

\item{threshold}{For 'seg' caa_method, how much of a chr-arm must be gained or lost to get -1 or 1 score}

\item{arm_ids}{[Default c("p", "q")]}

\item{verbose}{verbose statement}
}
\value{
A matrix of aneuploidy scores [chr-arms x caa_methods]
}
\description{
Takes the Arm and Seg level CN classification (Loss (-1),
Neut (0), Gain(1)) from the getCAA() function and reduces this data 
into a single matrix of chromosomal-arm level value (e.g. chr1p: 1)
}
