% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seg}
\alias{seg}
\title{Demo .seg file}
\format{A data frame with 384 ranges and 6 metadata columns variables:
\describe{
  \item{nMajor}{Haplotype A specific copy number}
  \item{nMinor}{Haplotype B specific copy number}
  \item{nAraw}{Haplotype A specific copy number (raw)}
  \item{nBraw}{Haplotype B specific copy number (raw)}
  \item{TCN}{Total copy number}
  \item{seg.mean}{Log 2 Ratio (L2R) segment mean}
  ...
}}
\source{
\url{https://portals.broadinstitute.org/ccle}
}
\usage{
seg
}
\description{
A .seg file from CCLE's 639-V urinary carcinoma cell line, processed using ASCAT
with the allele-specific and total copy numbers represented, as well as the L2R.
The data frame was converted into a GenomicRanges object.
}
\keyword{datasets}
