% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ucsc.mm9.cytoband}
\alias{ucsc.mm9.cytoband}
\title{mm9 UCSC cytoband table}
\format{A data frame with 403 rows and 5 variables:
\describe{
  \item{chrom}{Chromosome (e.g. chr1)}
  \item{chromStart}{Chromosome cytoband start position (0-based)}
  \item{chromEnd}{Chromosome cytoband end position (0-based)}
  \item{name}{Name of cytoband}
  \item{gieStain}{Giemsa staining}
  ...
}}
\source{
\url{https://genome.ucsc.edu/cgi-bin/}
}
\usage{
ucsc.mm9.cytoband
}
\description{
A data table containing the cytoband and centromere information
for mm9 genome
}
\keyword{datasets}
