% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkIfWGD.R
\name{checkIfWGD}
\alias{checkIfWGD}
\title{Whole genome doubling checker}
\usage{
checkIfWGD(segf, tcn_col, threshold = 0.5, wgd_gf = 0.5, wgd_ploidy = 2)
}
\arguments{
\item{segf}{.Seg Data [Data frame]}

\item{tcn_col}{Column ID for CN value [String]}

\item{threshold}{The threshold to call diploid
(Diploid = wgd_ploidy +/- threshold) (Default: 0.5) [Numeric]}

\item{wgd_gf}{Genomic fraction above wgd_ploidy
to be considered genome-doubled (Default: 0.5) [Numeric]}

\item{wgd_ploidy}{Base ploidy (Default: 2) [Integer]}
}
\value{
A named vector
 ploidy: The mean CN weighted by segment sizes
 WGD: Boolean indicated whether sample is WGD
}
\description{
Takes a .seg file data frame and estimates whether
the tumor is whole-genome doubled (WGD). This function requries
Total Copy Number and will not work using L2R.
}
