% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCAA.R
\name{getCAA}
\alias{getCAA}
\title{Seg to CN Chromosomal Fractions}
\usage{
getCAA(segf, cytoarm, tcn_col, filter_centromere = FALSE, ...)
}
\arguments{
\item{segf}{.Seg Data [Data frame]}

\item{cytoarm}{Output of cytobandToArm() function [List of Data frames]}

\item{tcn_col}{Column ID for CN value [String]}

\item{filter_centromere}{Whether to include or exclude segments
that overlap the centromere [Logical]}

\item{...}{Passes through:
.getChrarmFractions():  classify arm-CN into Loss/Neut/Gain [Boolean]
Parameters for .classifyCN() function (ploidy, threshold, verbose)}
}
\value{
A GRangesList split by chromosome with each
interval representing a reduced chromosomal aberration
and its associated chromosomal_arm and chromosome fraction
with and without NAs factored in
}
\description{
Takes a .seg data-frame and finds out for each CN
interval, what fraction of the chromosomal arm or chromosome does it
occupy. It calculates this metric across all covered regions in (NA-
excluded) or the entire chromosomal-arm/chromosome (NA-included). It 
also supports the Shukla et al. method of removing segments that 
overlap the centromere.
}
